Weapons updated.
Shotgun shoots smaller pellets
WepAttributes contains all the weapon's attributes. Easier to keep all
the values in one place.

Bullet now contains the script to send damage over to the zombie,
instead of the zombie calling for the damage. This is because
intuitively the bullet should know how much damage it is going to
transfer over to the zombie, not the other way round.

Added some tags for future use.