Move Speed Perk added.
Perk is currently added by pressing the 'm' key. 'm' for move, you see.

The perk multiplies movespeed by a factor of 1.25. I've yet to figure
out a good way to put the perk on the Inspector so that we can play with
the numbers, but for now the factor is hardcoded.

PerkInitializer also holds an int recording the number of skill points.
We can get the script to call zedResources for this number instead.

The good news is that this concept looks quite extendable to include
upgrades to weapon damage, etc.